Created "loadmodules" args for loading ".so" modules into EchoVault. Moved types associated with embedded commands extension into api_admin.go file and deleted types.go file as it's no longer necessary. Updated docker-compose and Dockerfile.dev to pass .so modules to load on startup. Volumes folder is no longer ignores except for the nodes subfolder.
